+++ ydb/core/tx/columnshard/engines/reader/common_reader/common/script_counters.cpp
@@ -1,0 +1,5 @@
+#include "script_counters.h"
+
+namespace NKikimr::NOlap::NReader::NCommon {
+
+}   // namespace NKikimr::NOlap::NReader::NCommon
