+++ ydb/core/tx/columnshard/engines/reader/common_reader/iterator/fetching.cpp
@@ -41,8 +41,8 @@ TConclusion<bool> TStepAction::DoExecuteImpl() {
     return FinishedFlag;
 }
 
-TStepAction::TStepAction(std::shared_ptr<IDataSource>&& source, TFetchingScriptCursor&& cursor, const NActors::TActorId& ownerActorId,
-    const bool changeSyncSection)
+TStepAction::TStepAction(
+    std::shared_ptr<IDataSource>&& source, TFetchingScriptCursor&& cursor, const NActors::TActorId& ownerActorId, const bool changeSyncSection)
     : TBase(ownerActorId, source->GetContext()->GetCommonContext()->GetCounters().GetAssembleTasksGuard())
     , Source(std::move(source))
     , Cursor(std::move(cursor)) {
@@ -50,177 +50,6 @@ TStepAction::TStepAction(std::shared_ptr<IDataSource>&& source, TFetchingScriptC
         Source->StartAsyncSection();
     } else {
         Source->CheckAsyncSection();
-    }
-}
-
-TConclusion<bool> TFetchingScriptCursor::Execute(const std::shared_ptr<IDataSource>& source) {
-    AFL_VERIFY(source);
-    NMiniKQL::TThrowingBindTerminator bind;
-    if (StepStartInstant == TMonotonic::Zero()) {
-        StepStartInstant = TMonotonic::Now();
-    }
-    Script->OnExecute();
-    AFL_VERIFY(!Script->IsFinished(CurrentStepIdx));
-    while (!Script->IsFinished(CurrentStepIdx)) {
-        if (source->HasStageData() && source->GetStageData().IsEmptyWithData()) {
-            source->OnEmptyStageData(source);
-            break;
-        } else if (source->HasStageResult() && source->GetStageResult().IsEmpty()) {
-            break;
-        }
-        auto step = Script->GetStep(CurrentStepIdx);
-        std::optional<TMemoryProfileGuard> mGuard;
-        if (IS_DEBUG_LOG_ENABLED(NKikimrServices::TX_COLUMNSHARD_SCAN_MEMORY)) {
-            mGuard.emplace("SCAN_PROFILE::FETCHING::" + step->GetName() + "::" + Script->GetBranchName());
-        }
-        AFL_DEBUG(NKikimrServices::TX_COLUMNSHARD_SCAN)("scan_step", step->DebugString())("scan_step_idx", CurrentStepIdx)(
-            "source_id", source->GetSourceId());
-
-        auto& schedulableTask = source->GetContext()->GetCommonContext()->GetSchedulableTask();
-        if (schedulableTask) {
-            schedulableTask->IncreaseExtraUsage();
-        }
-
-        const TMonotonic startInstant = TMonotonic::Now();
-        const TConclusion<bool> resultStep = step->ExecuteInplace(source, *this);
-        const auto executionTime = TMonotonic::Now() - startInstant;
-        source->GetContext()->GetCommonContext()->GetCounters().AddExecutionDuration(executionTime);
-
-        if (schedulableTask) {
-            schedulableTask->DecreaseExtraUsage(executionTime);
-        }
-
-        Script->AddStepDuration(CurrentStepIdx, executionTime, TMonotonic::Now() - StepStartInstant);
-        if (!resultStep) {
-            return resultStep;
-        }
-        if (!*resultStep) {
-            return false;
-        }
-        StepStartInstant = TMonotonic::Now();
-        ++CurrentStepIdx;
-    }
-    FOR_DEBUG_LOG(NKikimrServices::COLUMNSHARD_SCAN_EVLOG, source->AddEvent("fcursor"));
-    return true;
-}
-
-TString TFetchingScript::DebugString() const {
-    TStringBuilder sb;
-    TStringBuilder sbBranch;
-    for (auto&& i : Steps) {
-        sbBranch << "{" << i->DebugString() << "};";
-    }
-    if (!sbBranch) {
-        return "";
-    }
-    sb << "{branch:" << BranchName << ";steps:[" << sbBranch << "]}";
-    return sb;
-}
-
-TString TFetchingScript::ProfileDebugString() const {
-    TStringBuilder sb;
-    TStringBuilder sbBranch;
-    for (auto&& i : Steps) {
-        if (i->GetSumDuration() > TDuration::MilliSeconds(10)) {
-            sbBranch << "{" << i->DebugString(true) << "};";
-        }
-    }
-    if (!sbBranch) {
-        return "";
-    }
-    sb << "{branch:" << BranchName << ";";
-    if (AtomicGet(FinishInstant) && AtomicGet(StartInstant)) {
-        sb << "duration:" << AtomicGet(FinishInstant) - AtomicGet(StartInstant) << ";";
-    }
-
-    sb << "steps_10Ms:[" << sbBranch << "]}";
-    return sb;
-}
-
-void TFetchingScriptBuilder::AddAllocation(
-    const std::set<ui32>& entityIds, const NArrow::NSSA::IMemoryCalculationPolicy::EStage stage, const EMemType mType) {
-    if (Steps.size() == 0) {
-        AddStep(std::make_shared<TAllocateMemoryStep>(entityIds, mType, stage));
-    } else {
-        std::optional<ui32> addIndex;
-        for (i32 i = Steps.size() - 1; i >= 0; --i) {
-            if (auto allocation = std::dynamic_pointer_cast<TAllocateMemoryStep>(Steps[i])) {
-                if (allocation->GetStage() == stage) {
-                    allocation->AddAllocation(entityIds, mType);
-                    return;
-                } else {
-                    addIndex = i + 1;
-                }
-                break;
-            } else if (std::dynamic_pointer_cast<TAssemblerStep>(Steps[i])) {
-                continue;
-            } else if (std::dynamic_pointer_cast<TColumnBlobsFetchingStep>(Steps[i])) {
-                continue;
-            } else {
-                addIndex = i + 1;
-                break;
-            }
-        }
-        AFL_VERIFY(addIndex);
-        InsertStep<TAllocateMemoryStep>(*addIndex, entityIds, mType, stage);
-    }
-}
-
-TString IFetchingStep::DebugString(const bool stats) const {
-    TStringBuilder sb;
-    sb << "name=" << Name;
-    if (stats) {
-        sb << ";duration=" << GetSumDuration() << ";"
-           << "size=" << 1e-9 * GetSumSize();
-    }
-    sb << ";details={" << DoDebugString() << "};";
-    return sb;
-}
-
-TFetchingScriptBuilder::TFetchingScriptBuilder(const TSpecialReadContext& context)
-    : TFetchingScriptBuilder(context.GetReadMetadata()->GetResultSchema(), context.GetMergeColumns()) {
-}
-
-void TFetchingScriptBuilder::AddFetchingStep(const TColumnsSetIds& columns, const NArrow::NSSA::IMemoryCalculationPolicy::EStage stage) {
-    auto actualColumns = columns - AddedFetchingColumns;
-    AddedFetchingColumns += columns;
-    if (actualColumns.IsEmpty()) {
-        return;
-    }
-    if (Steps.size() && std::dynamic_pointer_cast<TColumnBlobsFetchingStep>(Steps.back())) {
-        TColumnsSetIds fetchingColumns = actualColumns + std::dynamic_pointer_cast<TColumnBlobsFetchingStep>(Steps.back())->GetColumns();
-        Steps.pop_back();
-        AddAllocation(actualColumns.GetColumnIds(), stage, EMemType::Blob);
-        AddStep(std::make_shared<TColumnBlobsFetchingStep>(fetchingColumns));
-    } else {
-        AddAllocation(actualColumns.GetColumnIds(), stage, EMemType::Blob);
-        AddStep(std::make_shared<TColumnBlobsFetchingStep>(actualColumns));
-    }
-}
-
-void TFetchingScriptBuilder::AddAssembleStep(
-    const TColumnsSetIds& columns, const TString& purposeId, const NArrow::NSSA::IMemoryCalculationPolicy::EStage stage, const bool sequential) {
-    auto actualColumns = columns - AddedAssembleColumns;
-    AddedAssembleColumns += columns;
-    if (actualColumns.IsEmpty()) {
-        return;
-    }
-    auto actualSet = std::make_shared<TColumnsSet>(actualColumns.GetColumnIds(), FullSchema);
-    if (sequential) {
-        const auto notSequentialColumnIds = GuaranteeNotOptional->Intersect(*actualSet);
-        if (notSequentialColumnIds.size()) {
-            AddAllocation(notSequentialColumnIds, stage, EMemType::Raw);
-            std::shared_ptr<TColumnsSet> cross = actualSet->BuildSamePtr(notSequentialColumnIds);
-            AddStep(std::make_shared<TAssemblerStep>(cross, purposeId));
-            *actualSet = *actualSet - *cross;
-        }
-        if (!actualSet->IsEmpty()) {
-            AddAllocation(actualSet->GetColumnIds(), stage, EMemType::RawSequential);
-            AddStep(std::make_shared<TOptionalAssemblerStep>(actualSet, purposeId));
-        }
-    } else {
-        AddAllocation(actualColumns.GetColumnIds(), stage, EMemType::Raw);
-        AddStep(std::make_shared<TAssemblerStep>(actualSet, purposeId));
     }
 }
