+++ ydb/core/tx/columnshard/engines/reader/common_reader/ya.make
@@ -6,6 +6,7 @@ SRCS(
 PEERDIR(
     ydb/core/tx/columnshard/engines/reader/common_reader/iterator
     ydb/core/tx/columnshard/engines/reader/common_reader/constructor
+    ydb/core/tx/columnshard/engines/reader/common_reader/common
 )
 
 END()
