+++ ydb/core/tx/columnshard/engines/reader/plain_reader/iterator/plain_read_data.h
@@ -5,7 +5,7 @@
 #include <ydb/core/tx/columnshard/engines/reader/abstract/read_context.h>
 #include <ydb/core/tx/columnshard/engines/reader/abstract/read_metadata.h>
 #include <ydb/core/tx/columnshard/engines/reader/common/queue.h>
-#include <ydb/core/tx/columnshard/engines/reader/common_reader/iterator/columns_set.h>
+#include <ydb/core/tx/columnshard/engines/reader/common_reader/common/columns_set.h>
 
 namespace NKikimr::NOlap::NReader::NPlain {
