+++ ydb/core/tx/columnshard/engines/reader/ya.make
@@ -13,6 +13,7 @@ PEERDIR(
     ydb/core/tx/program
     ydb/core/tx/columnshard/engines/reader/plain_reader
     ydb/core/tx/columnshard/engines/reader/simple_reader
+    ydb/core/tx/columnshard/engines/reader/common_reader
     ydb/core/tx/columnshard/engines/reader/abstract
     ydb/core/tx/columnshard/engines/reader/common
     ydb/core/tx/columnshard/engines/reader/actor
